// management/frontend/src/api/zod/system-menus/system-menus.ts
export const systemMenusReadMenusQuerySkipDefault = 0;
export const systemMenusReadMenusQuerySkipMin = 0;

export const systemMenusReadMenusQueryLimitDefault = 20;
export const systemMenusReadMenusQueryLimitMax = 500;



export const SystemMenusReadMenusQueryParams = zod.object({
  "skip": zod.number().min(systemMenusReadMenusQuerySkipMin).default(systemMenusReadMenusQuerySkipDefault),
  "limit": zod.number().max(systemMenusReadMenusQueryLimitMax).default(systemMenusReadMenusQueryLimitDefault),
  "name": zod.union([zod.string(),zod.null()]).optional(),
  "is_active": zod.union([zod.boolean(),zod.null()]).optional()
})